Improve performance of by() using NamedTuples
Remove GroupApplied and deprecate combine in favor of map(f, ::GroupedDataFrame). This avoids
storing a copy of the per-group data returned by the user-provided function. Take advantage of
this by allowing that function to return a NamedTuple. Introduce two completely different
code paths depending on whether the first returned object is DataFrame or a NamedTuple, as
the latter allows for more efficient operation by assuming that it represents a single row.
Use the same progressive eltype widening approach as Base.map so that we fill column vectors
whose types are known inside the kernel functions. This does not eliminate the type unstability
due to the fact that the user-provided function takes a DataFrame, but ensuring type stability for
half of the operations still improves performance significantly.

Also parameterize GroupedDataFrame on the type of data frame it wraps, and make its column index
have a concrete type. Deprecate an old map method for SubDataFrame. Fix a type unstability in hcat!.